Rename GetAllTournaments to GetTournaments throughout codebase
Refactored all references, files, classes, and namespaces from GetAllTournaments to GetTournaments for improved clarity and consistency. Updated API endpoints, application logic, infrastructure, and related tests to use the new naming convention.
