+++ src/BowlingMegabucks.TournamentManager.Application/Tournaments/GetTournaments/GetTournamentsQuery.cs
@@ -1,11 +1,11 @@
 using BowlingMegabucks.TournamentManager.Application.Abstractions.Messaging;
 
-namespace BowlingMegabucks.TournamentManager.Application.Tournaments.GetAllTournaments;
+namespace BowlingMegabucks.TournamentManager.Application.Tournaments.GetTournaments;
 
 /// <summary>
 /// Represents a query to retrieve all tournaments with offset-based pagination support.
 /// </summary>
-public sealed record GetAllTournamentsQuery
+public sealed record GetTournamentsQuery
     : IOffsetPaginationQuery<TournamentSummaryDto>
 {
     /// <inheritdoc />
