+++ src/BowlingMegabucks.TournamentManager.Application/Tournaments/GetTournaments/GetTournamentsQueryHandler.cs
@@ -2,25 +2,25 @@ using System.Diagnostics.CodeAnalysis;
 using BowlingMegabucks.TournamentManager.Application.Abstractions.Messaging;
 using ErrorOr;
 
-namespace BowlingMegabucks.TournamentManager.Application.Tournaments.GetAllTournaments;
+namespace BowlingMegabucks.TournamentManager.Application.Tournaments.GetTournaments;
 
 
 [SuppressMessage(
     "Performance",
     "CA1812:Avoid uninstantiated internal classes",
     Justification = "Instantiated by dependency injection container.")]
-internal sealed class GetAllTournamentsQueryHandler
-    : IOffsetPaginationQueryHandler<GetAllTournamentsQuery, TournamentSummaryDto>
+internal sealed class GetTournamentsQueryHandler
+    : IOffsetPaginationQueryHandler<GetTournamentsQuery, TournamentSummaryDto>
 {
     private readonly ITournamentQueries _tournamentQueries;
 
-    public GetAllTournamentsQueryHandler(ITournamentQueries tournamentQueries)
+    public GetTournamentsQueryHandler(ITournamentQueries tournamentQueries)
     {
         _tournamentQueries = tournamentQueries;
     }
 
     public async Task<ErrorOr<OffsetPaginationQueryResponse<TournamentSummaryDto>>> HandleAsync(
-        GetAllTournamentsQuery query,
+        GetTournamentsQuery query,
         CancellationToken cancellationToken)
     {
         IReadOnlyCollection<TournamentSummaryDto> results = await _tournamentQueries.GetAllTournamentsAsync(query, cancellationToken);
